Rename tokenPrices.ts to tokensPricesInCurrency.ts
